+++ e2e/wallets/eth-asset-context/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @coinmasters/eth-asset-context-test
 
+## 1.0.18
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.21
+
+## 1.0.17
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.20
+
 ## 1.0.16
 
 ### Patch Changes